refactor: rename conversion functions
