+++ opentelemetry-kotlin-compat/src/jvmMain/kotlin/io/embrace/opentelemetry/kotlin/creator/SpanContextCreatorImpl.kt
@@ -3,8 +3,8 @@ package io.embrace.opentelemetry.kotlin.creator
 import io.embrace.opentelemetry.kotlin.ExperimentalApi
 import io.embrace.opentelemetry.kotlin.aliases.OtelJavaSpanContext
 import io.embrace.opentelemetry.kotlin.tracing.SpanContextImpl
-import io.embrace.opentelemetry.kotlin.tracing.conversion.toOtelJavaTraceFlags
-import io.embrace.opentelemetry.kotlin.tracing.conversion.toOtelJavaTraceState
+import io.embrace.opentelemetry.kotlin.tracing.ext.toOtelJavaTraceFlags
+import io.embrace.opentelemetry.kotlin.tracing.ext.toOtelJavaTraceState
 import io.embrace.opentelemetry.kotlin.tracing.model.SpanContext
 import io.embrace.opentelemetry.kotlin.tracing.model.SpanContextAdapter
 import io.embrace.opentelemetry.kotlin.tracing.model.TraceFlags
