+++ opentelemetry-kotlin-compat/src/jvmMain/kotlin/io/embrace/opentelemetry/kotlin/tracing/export/OtelJavaSpanExporterExt.kt
