+++ opentelemetry-kotlin-compat/src/jvmMain/kotlin/io/embrace/opentelemetry/kotlin/tracing/ext/OtelJavaStatusDataExt.kt
@@ -1,12 +1,14 @@
-package io.embrace.opentelemetry.kotlin.tracing.conversion
+package io.embrace.opentelemetry.kotlin.tracing.ext
 
 import io.embrace.opentelemetry.kotlin.ExperimentalApi
 import io.embrace.opentelemetry.kotlin.aliases.OtelJavaStatusCode
+import io.embrace.opentelemetry.kotlin.aliases.OtelJavaStatusData
 import io.embrace.opentelemetry.kotlin.tracing.data.StatusData
 
 @OptIn(ExperimentalApi::class)
-internal fun OtelJavaStatusCode.toOtelKotlinStatusData(description: String?): StatusData = when (this) {
+internal fun OtelJavaStatusData.toOtelKotlinStatusData(): StatusData = when (statusCode) {
     OtelJavaStatusCode.UNSET -> StatusData.Unset
     OtelJavaStatusCode.OK -> StatusData.Ok
     OtelJavaStatusCode.ERROR -> StatusData.Error(description)
+    null -> StatusData.Unset
 }
