+++ opentelemetry-kotlin-compat/src/jvmTest/kotlin/io/embrace/opentelemetry/kotlin/tracing/export/OtelJavaSpanExporterAdapterTest.kt
@@ -5,8 +5,8 @@ import io.embrace.opentelemetry.kotlin.attributes.toMap
 import io.embrace.opentelemetry.kotlin.export.OperationResultCode
 import io.embrace.opentelemetry.kotlin.fakes.otel.java.FakeOtelJavaSpanExporter
 import io.embrace.opentelemetry.kotlin.fakes.otel.kotlin.FakeSpanData
-import io.embrace.opentelemetry.kotlin.tracing.conversion.toOtelJavaSpanKind
-import io.embrace.opentelemetry.kotlin.tracing.conversion.toOtelJavaStatusCode
+import io.embrace.opentelemetry.kotlin.tracing.ext.toOtelJavaSpanKind
+import io.embrace.opentelemetry.kotlin.tracing.ext.toOtelJavaStatusCode
 import org.junit.Before
 import org.junit.Test
 import kotlin.test.assertEquals
